Move Hasher and FastHasher from patch_subset/ to common/
