+++ common/fast_hasher.h
@@ -1,10 +1,10 @@
-#ifndef PATCH_SUBSET_FAST_HASHER_H_
-#define PATCH_SUBSET_FAST_HASHER_H_
+#ifndef COMMON_FAST_HASHER_H_
+#define COMMON_FAST_HASHER_H_
 
 #include "absl/strings/string_view.h"
-#include "patch_subset/hasher.h"
+#include "common/hasher.h"
 
-namespace patch_subset {
+namespace common {
 
 // Uses fast-hash (https://github.com/ztanml/fast-hash) to compute a checksum of
 // binary data.
@@ -16,6 +16,6 @@ class FastHasher : public Hasher {
   uint64_t Checksum(::absl::string_view data) const override;
 };
 
-}  // namespace patch_subset
+}  // namespace common
 
-#endif  // PATCH_SUBSET_FAST_HASHER_H_
+#endif  // COMMON_FAST_HASHER_H_
