+++ patch_subset/BUILD
@@ -97,7 +97,6 @@ cc_library(
     srcs = [
         "codepoint_map.cc",
         "compressed_set.cc",
-        "fast_hasher.cc",
         "integer_list_checksum_impl.cc",
         "vcdiff_binary_diff.cc",
     ],
@@ -105,8 +104,6 @@ cc_library(
         "codepoint_map.h",
         "compressed_set.h",
         "encodings.h",
-        "fast_hasher.h",
-        "hasher.h",
         "integer_list_checksum.h",
         "integer_list_checksum_impl.h",
         "patch_subset_server.h",
@@ -124,7 +121,6 @@ cc_library(
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/types:optional",
-        "@fasthash",
         "@harfbuzz",
         "@open-vcdiff//:vcdenc",
     ],
@@ -139,7 +135,6 @@ cc_test(
         "codepoint_map_test.cc",
         "compressed_set_test.cc",
         "fake_subsetter.h",
-        "fast_hasher_test.cc",
         "frequency_codepoint_predictor_test.cc",
         "harfbuzz_subsetter_test.cc",
         "integer_list_checksum_impl_test.cc",
