+++ patch_subset/client_server_integration_test.cc
@@ -1,12 +1,13 @@
 #include "common/brotli_binary_diff.h"
 #include "common/brotli_binary_patch.h"
+#include "common/fast_hasher.h"
 #include "common/file_font_provider.h"
 #include "common/font_data.h"
+#include "common/hasher.h"
 #include "common/hb_set_unique_ptr.h"
 #include "gtest/gtest.h"
 #include "patch_subset/codepoint_mapper.h"
 #include "patch_subset/compressed_set.h"
-#include "patch_subset/fast_hasher.h"
 #include "patch_subset/harfbuzz_subsetter.h"
 #include "patch_subset/integer_list_checksum.h"
 #include "patch_subset/integer_list_checksum_impl.h"
@@ -26,9 +27,11 @@ using common::BinaryDiff;
 using common::BinaryPatch;
 using common::BrotliBinaryDiff;
 using common::BrotliBinaryPatch;
+using common::FastHasher;
 using common::FileFontProvider;
 using common::FontData;
 using common::FontProvider;
+using common::Hasher;
 using common::hb_set_unique_ptr;
 using common::make_hb_set_from_ranges;
 using patch_subset::cbor::ClientState;
