+++ patch_subset/mock_hasher.h
@@ -2,13 +2,13 @@
 #define PATCH_SUBSET_MOCK_HASHER_H_
 
 #include "absl/strings/string_view.h"
+#include "common/hasher.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "patch_subset/hasher.h"
 
 namespace patch_subset {
 
-class MockHasher : public Hasher {
+class MockHasher : public common::Hasher {
  public:
   MOCK_METHOD(uint64_t, Checksum, (::absl::string_view data), (const override));
 };
