+++ patch_subset/patch_subset_client_test.cc
@@ -24,6 +24,7 @@ using common::BinaryPatch;
 using common::FileFontProvider;
 using common::FontData;
 using common::FontProvider;
+using common::Hasher;
 using common::hb_set_unique_ptr;
 using common::make_hb_set;
 using common::make_hb_set_from_ranges;
